Rename `*enabled*` goog-define (:closure-define) to `ENABLED`
Fixes a constant optimization error on goog-define *enabled*.
It is not possible to `set!` a goog-define.

This change only impacts an undocumented, advanced configuration feature:
Setting the `*enabled*` flag at build time from the cljs build config. Since the
cljs advanced compilation was broken before this fix, it is safe to assume no
one was using this feature. This commit must be listed in the next release’s
changelog.

*Migration guide:*

Only if you are already using this feature, in `shadow-cljs.edn`

```diff
- :closure-defines {hyperfiddle.rcf/*enabled* true|false}
+ :closure-defines {hyperfiddle.rcf/ENABLED   true|false}
```